requirements.txt
h5py==2.10.0 # 3.0.0 breaks keras
tensorflow
Keras
pandas
Pillow
sklearn